drivers/ukintctlr: enable PIC interrupt routing in APIC
Architecturally, the APIC at reset does not route the PIC INT line
to the CPU--this must be explicitly configured by setting the LVT_LINT0
register appropriately. By default, the in-kernel implementation in KVM
does configure it this way, but this non-standard behavior can be
disabled via a quirk flag, and some VMMs (such as OpenVMM) do this.

Support such VMMs by explicitly enabling PIC interrupt routing during
startup. This is harmless on qemu/fc.

This can be removed when the PIC is no longer used (by leveraging
the IOAPIC and/or MSIs for device interrupts).

Signed-off-by: John Starks <jostarks@microsoft.com>