lib/ukallocregion: Remove page-sized region check
Get rid of unnecessary initialization check for passed region being
bigger than one page.

Its reasoning was most likely related to the ukboot initialization
which is `addmem`-based. Whether the first free physical region
reported in the boot memory map is smaller than a page or not
has nothing to do with the boot specification and everything to do
with the platform and its firmware. There is no reason to inhibit
other region allocator users from initializing based on buffers
smaller than a page.

Nevertheless, this check has no place in the allocator library and
should be addressed by the caller, where the limitation lies.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>