arch: Add `__isr` attribute macro
In the case of some functions, we may want to be able to compile them
and be sure they do not use any extended registers (e.g., FPU, SIMD)
so that we can use them in an exception context or anywhere we do not
want to taint the said extended context.

To make it easy, add the `__isr` macro one can add to its function
to ensure it complies with the above.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>