lib/posix-socket: Poll for connection error on blocking sockets
On a blocking socket, whenever the driver would return -EINPROGRESS,
the core would attempt to poll for when the socket becomes available
for writing. If the other end however is no longer available then
the equivalent of an EPOLLHUP would be generated for which the
aforementioned polling does not check. This leads to the calling
thread blocking indefinetely.

Fix this by also polling for errors so that the event handler can
set the socket event and thus wake the blocked thread.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>