lib/nolibc: Fix return local variable warning
This change fixes a warning in getcwd about possibly returning a local
variable's address. The warning is technically true, as the caller
could, by sheer luck, pass buf == tmp which would cause a strdup.
This fix plugs that cornercase with a temporary local, preserving the
documented logic of calling strdup only if !buf.

Signed-off-by: Andrei Tatar <ttr@unikraft.io>