plat/xen/arm: Add `.headers` `PROGBITS` section
Since the PHDRs do not have a PROGBITS section associated with
the PT_LOAD segment that contains them and thus nothing to define their
memory layout placement relative to the other segments/sections,
the linker will place said segment at virtual/physical address 0x0.

Fix this by adding a dummy, non-empty, PROGBITS ".headers" section
to be part of this PHDR and whose memory placement is right before
the .text section. To maintain proper alignment of the .text section
that follows, pad this section to the next page-aligned boundary.

Lastly, for a bit of space efficiency reasons, ensure that the ELF
headers and this dummy padding are not a bigger than a page and
associate this dummy section with a meaningful 4-byte magic value
that is easy to find in hexdumps: 0xDEADB0B0 (we often use this for
magic values).

Checkpatch-Ignore: LONG_LINE
Signed-off-by: Sergiu Moga <sergiu@unikraft.io>