support/scripts: Adapt for PHDRs being part of first PT_LOAD segment
We place the Program Headers in a PT_LOAD ELF segment with the help of
a dummy PROGBITS 4-byte-sized section ".headers". However, the Linux
boot protocol on ARM64 involves no ELF loading so the ELF kernel is dumped
as a raw binary directly. For this to work, we use objcopy to convert
from ELF to raw binary. The way objcopy does this is with the help
of ELF Sections (not Program Headers). It goes through each one and
tries to write them into the newly converted file as if they were loaded
in memory.
Our problem is that we don't have a section associated with the contents
of the Program Headers themselves so objcopy will end up not copying them,
despite them being part of a loadable segment (it will only load
".headers"'s dummy 4 bytes). What objcopy will do is just copy
.headers to offset 0 of raw binary, followed by whatever padding there
is until .text.
Besides not having the Program Headers loaded in memory together with
the kernel, this also results in all offsets being off by
Elf64_Ehdr(.e_phoff) + Elf64_Ehdr.e_phnum * Elf64_Ehdr.e_phentsize
(assuming that e_phoff == sizeof(Elf64_Ehdr), i.e. Program Headers
immediately follow the ELF header).

To fix this, after writing the linux boot protocol header, simply append
the ELF headers that are loadable but do not have a PROGBITS section
associated with them, followed by our objcopy-created binary, since the
latter will contain the loadable bytes that would have come right after
the ELF headers we are interested in.

Checkpatch-Ignore: LONG_LINE
Signed-off-by: Sergiu Moga <sergiu@unikraft.io>