lib/posix-process/signal: Rework syserror handling
Rework handling of system errors so that the decision on whether to
signal the application is taken before jumping to the interrupt safe
context. This allows falling-through to the UNHANDLED_EXEPTION event
handler, if the signal is not deliverable.

Notice that although the current implementation could be updated to
call UK_CRASH_EX() with the application's context, falling-through
to the default handler is the right approach as it allows applying
a unified policy to unhandled system failures.

Checkpatch-Ignore: PREFER_FALLTHROUGH
Signed-off-by: Michalis Pappas <michalis@unikraft.io>