build/linker, plat/*: Move optional PDHR to last
The TLS program header can sometimes be a zero length LOAD segment. But
GRUB will complain if two LOAD program headers map to the same virtual
address, even if one is zero-length. The TLS header is the only header
which is likely to be both LOAD and zero-length, so by making it the
last LOAD header in the ELF file, this overlap will never occur.

Co-authored-by: Wyatt Borden <wyattborden@gmail.com>
Signed-off-by: Nathan Williams <nathanwilliams@utexas.edu>
Signed-off-by: Wyatt Borden <wyattborden@gmail.com>