lib/ukdebug: Make UK_BUG an unconditional halt
Previously UK_BUG() was a wrapper around UK_ASSERT(), thus allowing
execution to continue in builds with assertions disabled. This, however,
is not how it is used throughout the codebase, where it serves as an
explicit marker of an impossible code path that, if ever reached,
denotes a fundamental bug in the kernel.
This change makes UK_BUG unconditionally halt execution regardless of
configuration, and moves it to the <uk/crash.h> header.
Callers are unafected, as <uk/assert.h> includes <uk/crash.h> by
necessity.

Signed-off-by: Andrei Tatar <ttr@unikraft.io>